+++ anyblok_pyramid_beaker/config.py
@@ -1,6 +1,7 @@
 # This file is a part of the AnyBlok / Pyramid / Beaker project
 #
 #    Copyright (C) 2016 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -9,8 +10,8 @@ from anyblok.config import Configuration
 from anyblok_pyramid import config  # noqa for update config
 
 
-Configuration.applications['pyramid']['configuration_groups'].append('beaker')
-Configuration.applications['gunicorn']['configuration_groups'].append('beaker')
+Configuration.add_configuration_groups('pyramid', ['beaker'])
+Configuration.add_configuration_groups('gunicorn', ['beaker'])
 
 
 def get_db_name(request):